+++ addon/locales/en/translations.js
@@ -163,7 +163,8 @@ export default {
         'nempty': 'Not empty',
         'between': 'Between',
       },
-      'filter-condition': 'Condition'
+      'filter-condition': 'Condition',
+      'clear-filter-in-column': 'Clear the filter in this column',
     },
 
     'olv-filter-interval': {