+++ addon/locales/ru/translations.js
@@ -164,7 +164,8 @@ export default {
         'empty': 'Не заполнен',
         'between': 'В интервале',
       },
-      'filter-condition': 'Условие'
+      'filter-condition': 'Условие',
+      'clear-filter-in-column': 'Сбросить фильтр в этой колонке',
     },
 
     'olv-filter-interval': {